[ECO-5265] Refactored Discontinuities.kt to Discontinuity.kt 1. UpdatedHandlesDiscontinuity interface 2. Added RoomEvent as "room.discontinuity" to EventTypes
