+++ chat-android/src/main/java/com/ably/chat/EventTypes.kt
@@ -98,3 +98,15 @@ public enum class RoomReactionEventType(public val eventName: String) {
      */
     Reaction("roomReaction"),
 }
+
+/**
+ * Room events.
+ */
+public enum class RoomEvent(public val event: String) {
+    /**
+     * Event triggered when a discontinuity is detected in the room's channel connection.
+     * A discontinuity occurs when an attached or update event comes from the channel with resume=false,
+     * except for the first attach or attaches after explicit detach calls.
+     */
+    Discontinuity("room.discontinuity"),
+}
